+++ Functions/alliance.js
@@ -4,7 +4,7 @@ const inputValidation = require('./esiJS-Utils/inputValidation')
 module.exports = {
     /**
      * List all active player alliances.
-     * @async
+
      * @returns {number[]} A array of all active player alliances.
      */
     alliances() {
@@ -14,7 +14,7 @@ module.exports = {
     },
     /**
      * Get all current member alliances of an alliance.
-     * @async
+
      * @param ID {number} The alliance ID to get the alliances from.
      * @returns {number[]} The alliances in the alliance.
      */
@@ -31,7 +31,7 @@ module.exports = {
     },
     /**
      * Get the icon urls for a alliance.
-     * @async
+
      * @param allianceID {number} The alliance ID to get the icon of.
      * @returns {object} Links to the different sizes of the alliance icon.
      */
@@ -48,7 +48,7 @@ module.exports = {
     },
     /**
      * Get public information about an alliance.
-     * @async
+
      * @param ID {number} The alliance ID to get info from.
      * @returns {object} Public info on the alliance.
      */
@@ -67,7 +67,7 @@ module.exports = {
         /**
          * Get alliance contacts.
          * @param {number} allianceID
-         * @async
+
          * @requires esi-alliances.read_contacts.v1
          * @returns {JSON}
          */
@@ -86,7 +86,7 @@ module.exports = {
         /**
          * Get alliance contact labels
          * @param {number} allianceID
-         * @async
+
          * @requires esi-alliances.read_contacts.v1
          * @returns {JSON}
          */