+++ Functions/character.js
@@ -5,7 +5,7 @@ module.exports = {
     /**
      * Get public information about an alliance.
      * @exports info
-     * @async
+
      * @param characterID {number} The alliance ID to get info from.
      * @returns {JSON} Public info on the alliance.
      */
@@ -25,7 +25,7 @@ module.exports = {
     /**
      * GGet a list of all the corporations a character has been a member of.
      * @exports corpHistory
-     * @async
+
      * @param charID {number} The character to get the history of.
      * @returns {JSON} The character's history.
      */
@@ -43,7 +43,7 @@ module.exports = {
     /**
      * Get portrait urls for a character.
      * @exports portrait
-     * @async
+
      * @param charID {number} The character to get the portrait of.
      * @returns {JSON} Links to the different sizes of the character's portrait.
      */
@@ -61,7 +61,7 @@ module.exports = {
     /**
      * Get public information about a character.
      * @exports info
-     * @async
+
      * @param charID {number} The character to get the public info of.
      * @returns {JSON} Public info on a character.
      */
@@ -80,7 +80,7 @@ module.exports = {
      * Gets agents research.
      * @requires esi-characters.read_agents_research.v1
      * @param {number} characterID
-     * @async
+
 
      * @returns {JSON} A list of agents research information for a character. The formula for finding the current research points with an agent is: currentPoints = remainderPoints + pointsPerDay * days(currentTime - researchStartDate)
      */
@@ -100,7 +100,7 @@ module.exports = {
      * Get blueprints.
      * @param {number} characterID
      * @requires esi-characters.read_blueprints.v1
-     * @async
+
 
      * @returns {JSON} A list of blueprints the character owns.
      */
@@ -121,7 +121,7 @@ module.exports = {
      * @param {number} characterID
      * @param {[numbers]} characters The target characters to calculate the charge for.
      * @requires esi-characters.read_contacts.v1
-     * @async
+
      * @authencated
      * @returns {JSON} A CSPA charge cost.
      */
@@ -150,7 +150,7 @@ module.exports = {
      * Get jump fatigue.
      * @param {number} characterID
      * @requires esi-characters.read_fatigue.v1
-     * @async
+
 
      * @returns {JSON} A character’s jump activation and fatigue information.
      */
@@ -170,7 +170,7 @@ module.exports = {
      * Get medals.
      * @param {number} characterID
      * @requires esi-characters.read_medals.v1
-     * @async
+
 
      * @returns {JSON} A list of medals the character has.
      */
@@ -190,7 +190,7 @@ module.exports = {
      * Get character corporation roles.
      * @param {number} characterID
      * @requires esi-characters.read_corporation_roles.v1
-     * @async
+
 
      * @returns {JSON} A character’s corporation roles.
      */
@@ -210,7 +210,7 @@ module.exports = {
      * Get standings.
      * @param {number} characterID
      * @requires esi-characters.read_standings.v1
-     * @async
+
 
      * @returns {JSON} Character standings from agents, NPC corporations, and factions.
      */
@@ -230,7 +230,7 @@ module.exports = {
      * Yearly aggregate stats.
      * @param {number} characterID
      * @requires esi-characterstats.read.v1
-     * @async
+
 
      * @returns {JSON} Aggregate yearly stats for a character.
      */
@@ -250,7 +250,7 @@ module.exports = {
      * Get character corporation titles.
      * @param {number} characterID
      * @requires esi-characters.read_titles.v1
-     * @async
+
 
      * @returns {JSON} A character’s titles.
      */
@@ -271,7 +271,7 @@ module.exports = {
          * Get character assets.
          * @requires esi-assets.read_assets.v1
          * @param {number} characterID
-         * @async
+
 
         * @returns {JSON} A list of the characters assets.
         */
@@ -292,7 +292,7 @@ module.exports = {
          * @param {number} characterID
          * @param {number[]} itemIDs
          * @requires esi-assets.read_assets.v1
-         * @async
+
 
         * @returns {JSON} Locations for a set of item ids, which you can get from character assets endpoint. Coordinates for items in hangars or stations are set to (0,0,0)
         */
@@ -322,7 +322,7 @@ module.exports = {
          * @param {number} characterID
          * @param {number[]} itemIDs
          * @requires esi-assets.read_assets.v1
-         * @async
+
 
         * @returns {JSON} Names for a set of item ids, which you can get from character assets endpoint. Typically used for items that can customize names, like containers or ships.
         */
@@ -353,7 +353,7 @@ module.exports = {
         * List bookmarks
         * @param {number} characterID
         * @requires esi-bookmarks.read_character_bookmarks.v1
-        * @async
+
 
         * @returns {JSON} A list of your character’s personal bookmarks.
         */
@@ -373,7 +373,7 @@ module.exports = {
          * List bookmark folders
          * @param {number} characterID
          * @requires esi-bookmarks.read_character_bookmarks.v1
-         * @async
+
 
         * @returns {JSON} A list of your character’s personal bookmark folders.
         */
@@ -397,7 +397,7 @@ module.exports = {
          * @param {number} characterID
          * @param {number} fromEvent OPTIONAL - The event ID to start retrieving events from.
          * @requires esi-calendar.read_calendar_events.v1
-         * @async
+
          * @returns {JSON} Get 50 event summaries from the calendar. If no from_event ID is given, the resource will return the next 50 chronological event summaries from now. If a from_event ID is specified, it will return the next 50 chronological event summaries from after that event
          */
         calendar(characterID, fromEvent) {
@@ -426,7 +426,7 @@ module.exports = {
          * @param {number} characterID
          * @param {number} eventID
          * @requires esi-calendar.read_calendar_events.v1
-         * @async
+
          * @returns {JSON} Get all the information for a specific event
          */
         getEvent(characterID, eventID) {
@@ -501,7 +501,7 @@ module.exports = {
          * Get clones.
          * @param {number} characterID
          * @requires esi-clones.read_clones.v1
-         * @async
+
          * @returns {JSON} A list of the character’s clones.
          */
         clones(characterID) {
@@ -520,7 +520,7 @@ module.exports = {
          * Get active implants.
          * @param {number} characterID
          * @requires esi-clones.read_implants.v1
-         * @async
+
          * @returns {JSON} Implants on the active clone of a character.
          */
         implants(characterID) {
@@ -540,7 +540,7 @@ module.exports = {
         /**
          * Get contacts
          * @param {number} characterID
-         * @async
+
          * @requires esi-characters.read_contacts.v1
          * @returns {JSON}
          */
@@ -560,7 +560,7 @@ module.exports = {
          * Bulk add contacts with same settings
          * @param {number} characterID
          * @param {number[]} contacts
-         * @async
+
          * @requires esi-characters.write_contacts.v1
          * @returns {JSON}
          */
@@ -586,7 +586,7 @@ module.exports = {
          * Bulk delete contacts
          * @param {number} characterID
          * @param {number[]} contacts
-         * @async
+
          * @requires esi-characters.write_contacts.v1
          * @returns {JSON}
          */
@@ -612,7 +612,7 @@ module.exports = {
          * Bulk edit contacts with same settings
          * @param {number} characterID
          * @param {number[]} contacts
-         * @async
+
          * @requires esi-characters.write_contacts.v1
          * @returns {JSON}
          */
@@ -656,8 +656,8 @@ module.exports = {
         /**
          * Get contract bids
          * @requires esi-contracts.read_character_contracts.v1
-         * @param {number} characterID 
-         * @param {number} contractID 
+         * @param {number} characterID
+         * @param {number} contractID
          * @returns bids on a particular auction contract.
          */
         bids(characterID, contractID) {
@@ -679,8 +679,8 @@ module.exports = {
         /**
          * Get contract items
          * @requires esi-contracts.read_character_contracts.v1
-         * @param {number} characterID 
-         * @param {number} contractID 
+         * @param {number} characterID
+         * @param {number} contractID
          * @returns items of a particular contract
          */
         items(characterID, contractID) {
@@ -705,7 +705,7 @@ module.exports = {
          * List character industry jobs
          * @param {number} characterID
          * @requires esi-industry.read_character_jobs.v1
-         * @async
+
          * @returns {JSON}
          */
         jobs(characterID) {
@@ -726,7 +726,7 @@ module.exports = {
           * Get character notifications.
           * @param {number} characterID The character to get the notifications of.
           * @requires esi-characters.read_notifications.v1
-          * @async
+
 
          * @returns {JSON} Character notifications.
          */
@@ -746,7 +746,7 @@ module.exports = {
          * Get new contact notifications.
          * @param {number} characterID
          * @requires esi-characters.read_notifications.v1
-         * @async
+
 
         * @returns {JSON} Notifications about having been added to someone’s contact list.
         */