+++ Functions/incursions.js
@@ -3,10 +3,12 @@ const request = require('./esiJS-Utils/request')
 module.exports = {
     /**
      * Return a list of current incursions.
-     * @async
+
      * @returns {object}
      */
-    incursions () {
-        return request({ subUrl: `incursions` })
+    incursions() {
+        return request({
+            subUrl: `incursions`
+        })
     }
 }